Fix some physics things somewhat before I try to do a full physics refactor
Get block pushing back to the same broken state as before, fix grounded gravity, fix bouncing on enemy heads and fix snake and caveman collision. Next I want to try and refactor everything so that velocity is only updated from one place so that it's easy to track and set the final velocity for an entity, fix the collision info struct or at least how we interact with it because currently we interact with it from everywhere and I think it's easy to get unforseen results, make it so that both players, enemies, bombs and blocks inherits from or uses the same components, fix the block pushing and collision issues we have now and in general just make everything easier to understand and less error prone. Here's hoping.