Make it so the spider is aware of which block he's hanging from
So that he can fall if that block is destroyed. Also fix the jumping animation for spiders, but that was fixed in my sprite animator package, which apparently is not tracked here at all. That seems wrong. Maybe I need to make releases on github so that my project can target a particular release? The ClimbingState changes I'm not sure of. I was messing around with the SpriteAnimator package to fix the spider jumping and climbing was broken and fixed, but it's currently working so I'm leaving it as is. I will refactor the SpriteAnimator later on and then everything will be affected regardless. fix #21 fix #26